user_terminal/requirements.txt
streamlit
pandas
numpy
streamlit_code_editor
PyGithub
streamlit-authenticator
st-files-connection
streamlit-autorefresh
matplotlib